backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2025.8.3
cffi==1.17.0
charset-normalizer==3.4.0
click==8.2.1
cohere==5.18.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
executing==2.1.0
fastapi==0.116.1
fastavro==1.12.1
fastjsonschema==2.20.0
filelock==3.16.0
fonttools==4.53.1
fqdn==1.5.1
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.35.3
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.29.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.2
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.2
notebook_shim==0.2.4
numpy==2.3.3
overrides==7.7.0
packaging==23.2
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pipenv==2025.0.2
platformdirs==4.3.3
pluggy==1.4.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
Pygments==2.18.0
pyparsing==3.1.4
pyperclip==1.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==2.0.7
python-multipart==0.0.20
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
scikit-learn==1.7.2
scipy==1.16.2
Send2Trash==1.8.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.43
sqlmodel==0.0.24
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.3
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2024.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.35.0
virtualenv==20.26.4
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1

# --- Optional platform-specific packages ---
windows-curses==2.3.2; platform_system == "Windows"
pywinpty==2.0.14; platform_system == "Windows"
pynput==1.7.6; sys_platform == "win32"

# --- Linux-safe replacement for evdev ---
evdev-binary; platform_system == "Linux"

